Moved booked appointments logic to correct component
